+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 1.17.0.dev
 ==========
 
+* enhancement: Specify tasks from which training/tuning operator to transform/deploy in related operators
 * bug-fix: Revert appending Airflow retry id to default job name
 * bug-fix: Session: don't allow get_execution_role() to return an ARN that's not a role but has "role" in the name
 * bug-fix: Remove ``__all__`` from ``__init__.py`` files